chore(notices) Extracted some functionality to notice.ts to satisfy dependency constraints
