+++ editor/src/components/editor/action-types.ts
@@ -48,7 +48,7 @@ import {
   PersistentModel,
   StoredEditorState,
 } from './store/editor-state'
-import { Notice } from '../common/notices'
+import { Notice } from '../common/notice'
 import { BuildType } from '../../core/workers/ts/ts-worker'
 import type { EditorTab } from './store/editor-tabs'
 export { isLoggedIn, loggedInUser, LoginState, notLoggedIn, UserDetails } from '../../common/user'
