+++ editor/src/components/editor/actions/action-creators.ts
@@ -46,7 +46,7 @@ import type { RightMenuTab } from '../../canvas/right-menu'
 import type { CodeEditorTheme } from '../../code-editor/code-editor-themes'
 import type { CursorPosition } from '../../code-editor/code-editor-utils'
 import type { EditorPane, EditorPanel } from '../../common/actions'
-import type { Notice } from '../../common/notices'
+import { Notice } from '../../common/notice'
 import type { CodeResultCache, PropertyControlsInfo } from '../../custom-code/code-file'
 import type { ElementContextMenuInstance } from '../../element-context-menu'
 import type { FontSettings } from '../../inspector/common/css-utils'
